+++ mobile/cmp/navigator/impl/Page.ts
@@ -10,6 +10,7 @@ import {elementFromContent} from '@xh/hoist/utils/react';
 import {useRef} from 'react';
 import {PageModel} from '../PageModel';
 import './Page.scss';
+import {errorBoundary} from '@xh/hoist/cmp/error/ErrorBoundary';
 
 /**
  * Wrapper for contents to be shown within a Navigator. This Component is used by Navigator's
@@ -25,7 +26,7 @@ export const page = hoistCmp.factory({
     model: uses(PageModel, {publishMode: 'limited'}),
 
     render({model}) {
-        const {content, props, isActive, renderMode, refreshContextModel} = model,
+        const {content, props, isActive, renderMode, refreshContextModel, errorBoundaryModel} = model,
             wasActivated = useRef(false);
 
         if (!wasActivated.current && isActive) wasActivated.current = true;
@@ -42,7 +43,10 @@ export const page = hoistCmp.factory({
             model: refreshContextModel,
             item: onsenPage({
                 className: 'xh-page',
-                item: elementFromContent(content, props)
+                item: errorBoundary({
+                    model: errorBoundaryModel,
+                    item: elementFromContent(content, props)
+                })
             })
         });
     }